Refactor SpotifyPolling services and fix duplicated types/mocks
- Refactor `SpotifyPlayerManager` to use functional state updates for concurrency safety.
- Centralize `SpotifyTokenResponse` in `types/spotify.ts` to remove duplication.
- Update `SpotifyPolling.startPolling` to use `isReady()` check and log warnings.
- Consolidate test mocks into `tests/unit/spotify-mocks.ts` and use `require` in `jest.mock` to resolve hoisting issues.
- Remove redundant duplicate code in test files.
