+++ types/spotify.ts
@@ -45,3 +45,11 @@ export interface UserPlaylistDto {
   owner: string
   public: boolean
 }
+
+export interface SpotifyTokenResponse {
+  access_token: string
+  token_type: string
+  scope: string
+  expires_in: number
+  refresh_token?: string
+}
